# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HR Management App API Documentation",
        "contact": {
            "email": "ammarelw@gmail.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/manageCollaborator/deleteEvaluation/{eval_id}": {
            "delete": {
                "tags": [
                    "evaluations"
                ],
                "summary": "Delete a specified evaluation from storage",
                "description": "Delete a specified evaluation from storage",
                "operationId": "App\\Http\\Controllers\\EvaluationController::destroy",
                "requestBody": {
                    "$ref": "#/components/requestBodies/evaluationsRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
...
        "/api/manageCollaborator/deleteLeave/{leave_id}": {
            "delete": {
                "tags": [
                    "Leaves"
                ],
                "summary": "Delete a specified leave from storage",
                "description": "Delete a specified leave from storage",
                "operationId": "App\\Http\\Controllers\\LeaveController::destroy",
                "requestBody": {
                    "$ref": "#/components/requestBodies/leaveRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
...
        "/api/manageCollaborator/deleteSkill/{skill_id}": {
            "delete": {
                "tags": [
                    "Skills"
                ],
                "summary": "Revoke a skill from a user",
                "description": "Revoke a skill to a user and delete it if no user still have it",
                "operationId": "App\\Http\\Controllers\\SkillController::destroy",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "evaluation": {
                "title": "evaluation",
                "description": "Evaluation object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "Monthly test"
                    },
                    "evalDate": {
                        "type": "string",
                        "format": "dateTime",
                        "example": "2020-10-05"
                    },
                    "result": {
                        "type": "string",
                        "example": "passed"
                    }
                },
                "type": "object"
            },
            "leave": {
                "title": "Leave",
                "description": "Leave object",
                "properties": {
                    "startingFromDate": {
                        "type": "string",
                        "format": "dateTime",
                        "example": "2020-10-05"
                    },
                    "nbDays": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "success": {
                "description": "Successful operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                }
            },
            "unauthorized": {
                "description": "Forbidden",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "This action is unauthorized."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "unauthenticated": {
                "description": "Authentication is required",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
...
        "requestBodies": {
...
            "evaluationsRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/evaluation"
                        }
                    }
                }
            },
            "leaveRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/leave"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}